refactor `@web-builder/styled` module
