+++ packages/builder-web-styled-components/core/index.ts
@@ -1,0 +1,7 @@
+export { StyledComponentDeclaration } from "./styled-component-declaration";
+export { composeStyledComponentVariableDeclaration } from "./compose-styled-component-declaration";
+export type {
+  StyledComponentJSXElementConfig,
+  NoStyleJSXElementConfig,
+  NamePreference,
+} from "./types";
