+++ packages/builder-web-styled-components/index.ts
@@ -1,3 +1,3 @@
 export * from "./main";
-export * from "./styled-component-declaration";
+export * from "./core";
 export * from "./optimization-post";
