+++ packages/builder-web-styled-components/main.ts
@@ -1,27 +1,13 @@
 import { handle } from "@coli.codes/builder";
-import { CSSProperties } from "@coli.codes/css";
 import { ScopedVariableNamer } from "@coli.codes/naming";
 import { WidgetWithStyle } from "@web-builder/core";
-import { JSXAttributes, JSXIdentifier } from "coli";
+import { JSXIdentifier } from "coli";
 import {
   composeStyledComponentVariableDeclaration,
   NamePreference,
-  StyledComponentDeclaration,
-} from "./styled-component-declaration";
-
-export interface StyledComponentJSXElementConfig {
-  id: string;
-  tag: JSXIdentifier;
-  attributes?: JSXAttributes;
-  style: CSSProperties;
-  styledComponent: StyledComponentDeclaration;
-}
-
-export interface NoStyleJSXElementConfig {
-  id?: string;
-  tag: JSXIdentifier;
-  attributes?: JSXAttributes;
-}
+  NoStyleJSXElementConfig,
+  StyledComponentJSXElementConfig,
+} from "./core";
 
 /**
  *
